+++ lib/src/material/containers/colored_box.dart
@@ -1,0 +1,65 @@
+import 'package:flutter/material.dart';
+
+import 'package:exui/src/widget/containers/colored_box.dart';
+
+extension BoxColorsExtension on Widget {
+  Widget redBox() => coloredBox(Colors.red);
+
+  Widget redAccentBox() => coloredBox(Colors.redAccent);
+
+  Widget greenBox() => coloredBox(Colors.green);
+
+  Widget greenAccentBox() => coloredBox(Colors.greenAccent);
+
+  Widget blueBox() => coloredBox(Colors.blue);
+
+  Widget blueAccentBox() => coloredBox(Colors.blueAccent);
+
+  Widget yellowBox() => coloredBox(Colors.yellow);
+
+  Widget yellowAccentBox() => coloredBox(Colors.yellowAccent);
+
+  Widget orangeBox() => coloredBox(Colors.orange);
+
+  Widget orangeAccentBox() => coloredBox(Colors.orangeAccent);
+
+  Widget purpleBox() => coloredBox(Colors.purple);
+
+  Widget purpleAccentBox() => coloredBox(Colors.purpleAccent);
+
+  Widget deepPurpleBox() => coloredBox(Colors.deepPurple);
+
+  Widget deepPurpleAccentBox() => coloredBox(Colors.deepPurpleAccent);
+
+  Widget pinkBox() => coloredBox(Colors.pink);
+
+  Widget pinkAccentBox() => coloredBox(Colors.pinkAccent);
+
+  Widget brownBox() => coloredBox(Colors.brown);
+
+  Widget tealBox() => coloredBox(Colors.teal);
+
+  Widget tealAccentBox() => coloredBox(Colors.tealAccent);
+
+  Widget cyanBox() => coloredBox(Colors.cyan);
+
+  Widget cyanAccentBox() => coloredBox(Colors.cyanAccent);
+
+  Widget lightBlueBox() => coloredBox(Colors.lightBlue);
+
+  Widget lightBlueAccentBox() => coloredBox(Colors.lightBlueAccent);
+
+  Widget lightGreenBox() => coloredBox(Colors.lightGreen);
+
+  Widget lightGreenAccentBox() => coloredBox(Colors.lightGreenAccent);
+
+  Widget limeBox() => coloredBox(Colors.lime);
+
+  Widget limeAccentBox() => coloredBox(Colors.limeAccent);
+
+  Widget greyBox() => coloredBox(Colors.grey);
+
+  Widget blackBox() => coloredBox(Colors.black);
+
+  Widget whiteBox() => coloredBox(Colors.white);
+}